Fix modal resolution hook naming and shape
